+++ Pandora/Paradigm/Primary/Transformer/Yoneda.hs
@@ -5,7 +5,7 @@ module Pandora.Paradigm.Primary.Transformer.Yoneda where
 import Pandora.Pattern.Semigroupoid ((.))
 import Pandora.Pattern.Category ((<--))
 import Pandora.Pattern.Functor.Covariant (Covariant ((<-|-)))
-import Pandora.Pattern.Transformer.Liftable (Liftable (lift))
+import Pandora.Pattern.Transformation.Liftable (Liftable (lift))
 import Pandora.Paradigm.Algebraic.Exponential ()
 
 newtype Yoneda t a = Yoneda