+++ Pandora/Pattern.hs
@@ -3,7 +3,7 @@ module Pandora.Pattern (module Exports) where
 
 import Pandora.Pattern.Betwixt as Exports
 import Pandora.Pattern.Object as Exports
-import Pandora.Pattern.Transformer as Exports
+import Pandora.Pattern.Transformation as Exports
 import Pandora.Pattern.Functor as Exports
 import Pandora.Pattern.Morphism as Exports
 import Pandora.Pattern.Groupoid as Exports