+++ Pandora/Pattern/Transformer.hs
@@ -1,5 +1,0 @@
-module Pandora.Pattern.Transformer (module Exports) where
-
-import Pandora.Pattern.Transformer.Lowerable as Exports
-import Pandora.Pattern.Transformer.Liftable as Exports
-import Pandora.Pattern.Transformer.Hoistable as Exports